Working on the local gaussian copula algorithms
